+++ packages/nhost-js/CHANGELOG.md
@@ -1,3 +1,5 @@
+- fix (packages/nhost-js): force refresh correctly if marginSeconds is set to 0 #54
+
 # 5.0.0-beta.9
 
 - fix (packages/nhost-js): if session doesn't have decodedToken treat as expired #49